增加调试信息位数

+++ driver/inc/Driver_Magic.h
@@ -8,7 +8,7 @@
 
 typedef struct {
     uint8_t  buf[MAGIC_MAX_LENGTH];
-    uint16_t sta;
+    uint32_t sta;
 
     char *name;
     int   defaultValue; // 默认值
@@ -16,14 +16,14 @@ typedef struct {
 } MagicHandle_Type;
 
 typedef struct {
-    int16_t debug1;
-    int16_t debug2;
-    int16_t debug3;
-    int16_t debug4;
-    int16_t debug5;
-    int16_t debug6;
-    int16_t debug7;
-    int16_t debug8;
+    int32_t debug1;
+    int32_t debug2;
+    int32_t debug3;
+    int32_t debug4;
+    int32_t debug5;
+    int32_t debug6;
+    int32_t debug7;
+    int32_t debug8;
 } DebugData_Type;
 
 /**
